+++ MoodHolicBE/src/main/java/akatsuki/moodholic/controller/CalendarController.java
@@ -2,6 +2,7 @@ package akatsuki.moodholic.controller;
 
 import akatsuki.moodholic.domain.Diary;
 import akatsuki.moodholic.dto.Calendar;
+import akatsuki.moodholic.service.CalendarFacadeService;
 import akatsuki.moodholic.service.CalendarService;
 import akatsuki.moodholic.service.DiaryService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -16,32 +17,26 @@ import java.util.List;
 @RequestMapping("/calendar")
 @Tag(name = "캘린더 관련 컨트롤러", description = "달력 관련 로직을 처리하기 위한 컨트롤러 입니다.")
 public class CalendarController {
-    private final CalendarService calendarService;
-    private final DiaryService diaryService;
+
+    private final CalendarFacadeService calendarFacadeService;
 
     @Autowired
-    public CalendarController(CalendarService calendarService, DiaryService diaryService) {
-        this.calendarService = calendarService;
-        this.diaryService = diaryService;
+    public CalendarController(CalendarFacadeService calendarFacadeService) {
+        this.calendarFacadeService = calendarFacadeService;
     }
 
     @GetMapping("/{memberId}")
     @Operation(summary = "달력 리스트 출력", description = "맴버가 작성한 기록이 있는 모든 다이어리 날짜와 저장상태들을 반환합니다." +
             "\n만약 임시저장되어 있는 다이어리도 있을 수 있기 때문에 상태값도 포함하여 반환합니다.")
     public ResponseEntity<List<Calendar>> getCalendar(@PathVariable long memberId){
-        List<Calendar> returnValue = calendarService.getCalendar(getDiaries(memberId));
+        List<Calendar> returnValue = calendarFacadeService.getCalendar(memberId);
         return ResponseEntity.ok().body(returnValue);
     }
 
     @GetMapping("/year/{memberId}")
     public ResponseEntity<List<Calendar>> getCalendarOfYear(@PathVariable long memberId, @RequestParam(name = "year") int year){
-        List<Calendar> returnValue = calendarService.getCalendarOfYear(getDiaries(memberId),year);
+        List<Calendar> returnValue = calendarFacadeService.getCalendarOfYear(memberId,year);
         return ResponseEntity.ok().body(returnValue);
-    }
-
-    private List<Diary> getDiaries(long memberId) {
-        List<Diary> diaryList = diaryService.getMemberDiaries(memberId);
-        return diaryList;
     }
 
 }
